refactor(api): phase 2 schema splits and executor helpers
