+++ api/chat/tools/schemas/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Zod schemas for chat tool input validation
+ *
+ * Domain modules are split under ./schemas/*.ts; this barrel re-exports all.
+ */
+
+export * from "./memory.js";
+export * from "./songs.js";
+export * from "./documents.js";
+export * from "./cursor.js";
+export * from "./maps.js";
+export * from "./media.js";
+export * from "./app-state.js";
